% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{f}
\alias{f}
\title{A laplace function}
\usage{
f(x)
}
\arguments{
\item{x}{the independent value}
}
\value{
the value of the function
}
\description{
A probability density function of laplace function
}
