% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StatComp20030R.R
\name{laplacerandnum}
\alias{laplacerandnum}
\title{Random number of Laplacian distribution using R}
\usage{
laplacerandnum(n, beta, mu)
}
\arguments{
\item{n}{the number of samples}

\item{beta}{location parameter of Laplacian distribution}

\item{mu}{the scale parameter of Laplacian distribution}
}
\value{
a random sample of size \code{n}
}
\description{
Random number of Laplacian distribution using R
}
\examples{
\dontrun{
x<-laplacerandnum(1e4,3,0)
hist(x, freq = FALSE,main = "Histogram with the density curve", xlab = "value")
}
}
