% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rwMetropolis}
\alias{rwMetropolis}
\title{A random walk sampling using Rcpp}
\usage{
rwMetropolis(sigma, x0, N)
}
\arguments{
\item{sigma}{the parameter of variance}

\item{x0}{the initial value}

\item{N}{the number of uniform distributed random samples}
}
\value{
a random sample of size \code{n}
}
\description{
A random walk sampling using Rcpp
}
