% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StatComp20030R.R
\name{sigma2test}
\alias{sigma2test}
\title{A confidence interval and hypothesis test using R}
\usage{
sigma2test(x, mu, sigma, hypothesis, a)
}
\arguments{
\item{x}{samples}

\item{mu}{the mean value of the normal distribution}

\item{sigma}{the variance of the normal distribution}

\item{hypothesis}{inspection required}

\item{a}{confidence level}
}
\value{
acceptance or rejection and the confidence interval of estimation
}
\description{
A confidence interval and hypothesis test of single normal variance using R
}
\examples{
\dontrun{
A<-rnorm(100,0,2)
rnR<-sigma.test(A,mu=0,sigma=2,hypothesis='equal',a=0.05)
rnR
}
}
