\name{StatComp20030-package}
\alias{StatComp20030-package}
\alias{StatComp20030}
\docType{package}
\title{
  Hypothesis test and confidence interval and random numbers
}
\description{
  Hypothesis test and confidence interval of variance of nonormal population and Generation of Laplacian random numbers.
}
\details{
  This section should provide a more detailed overview of how to use the
  package, including the most important functions.
}
\author{
Li Wenyi, 1136808378@qq.con

Maintainer: Li Wenyi <1136808378@qq.com>
}
\references{
  This optional section can contain literature or other references for
  background information.
}
\keyword{ package }
\seealso{
  Optional links to other man pages
}
\examples{
  \dontrun{
     ## Optional simple examples of the most important functions
     ## These can be in \dontrun{} and \donttest{} blocks.   
  }
}
